move sdl2 include dir

+++ src/include/GameConfig.h
@@ -13,8 +13,8 @@ class GameConfig {
         Uint16 view_height;
         Uint16 view_width;
 
-        Uint8 player_speed_forward;
-        Uint8 player_speed_side;
+        float player_speed_forward;
+        float player_speed_side;
 
         GameConfig();
 };
